+++ lib/cyrax/base.rb
@@ -9,4 +9,20 @@ class Cyrax::Base
     @accessor = options[:as]
     @params = options[:params]
   end
+
+  def has_extension?(name)
+    self.class.has_extension?(name)
+  end
+
+  class << self
+    def register_extension(name)
+      @_extensions ||= []
+      @_extensions << name.to_sym
+    end
+
+    def has_extension?(name)
+      @_extensions ||= []
+      @_extensions.include?(name.to_sym)
+    end
+  end
 end