+++ lib/cyrax/extensions/has_service.rb
@@ -2,6 +2,10 @@ module Cyrax::Extensions
   module HasService
     extend ActiveSupport::Concern
 
+    included do
+      register_extension :has_service
+    end
+
     def collection
       respond_with build_collection, name: collection_name, present: :collection
     end