+++ spec/cyrax/extensions/has_decorator_spec.rb
@@ -12,38 +12,39 @@ class Foo;
 end
 
 module Cyrax
-  describe Cyrax::Extensions::HasDecorator do
+  class BaseWithDecorator < Cyrax::Base
     include Cyrax::Extensions::HasDecorator
+  end
+
+  describe Cyrax::Extensions::HasDecorator do
+    subject { BaseWithDecorator.new }
 
     describe 'class attributes' do
-      subject { self.class }
-      it { should respond_to(:decorator_class_name) }
+      its(:class) { should respond_to(:decorator_class_name) }
     end
 
     describe 'class methods' do
       describe '#decorator' do
-        before { self.class.decorator(:foo) }
-        subject { self }
+        before { subject.class.decorator(:foo) }
 
         its(:decorator_class_name) { should eq('foo') }
       end
     end
 
     describe 'instance methods' do
-      subject { self }
-
       describe '#decorator_class' do
-        before { self.class.decorator(:foo) }
+        before { subject.class.decorator(:foo) }
         its(:decorator_class) { should eq(Foo) }
       end
 
       describe '#decorable?' do
         context 'when `decorator_class_name` present' do
-          before { self.class.decorator(:foo) }
+          before { subject.class.decorator(:foo) }
           its(:decorable?) { should be_true }
         end
 
         context 'when `decorator_class_name` empty' do
+          before { subject.class.decorator(nil) }
           its(:decorable?) { should be_false }
         end
       end