+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+=== 0.3.0
+* Remove `wrapped_collection` from has_resource
+* Ability to pass Cyrax::Response to `respond_with`
+* Fix respond_with on Cyrax::Base
+
 === 0.2.3
 * Added collection decorator, to decorate collections with decorator use Decorator.decorate_collection(collection)
 * Added presenter logic